There are two much play calls
On testing the player we noticed that there are many play events so we tried to figure out when and where this are happening. In our case we don't need this two mse.play calls. Videos autplay or Players.play() does anything we need. So we think this calls are obsolete